web/web-backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
elasticsearch>=8.11.0
python-dotenv>=1.0.0
requests>=2.31.0
sse-starlette>=1.8.2
